+++ cmf-cli/Commands/new/TestCommand.cs
@@ -1,14 +1,14 @@
-using System;
-using System.Collections.Generic;
-using System.CommandLine;
-using System.CommandLine.NamingConventionBinder;
-using System.IO.Abstractions;
-using System.Text.Json;
 using Cmf.CLI.Constants;
 using Cmf.CLI.Core.Attributes;
 using Cmf.CLI.Core.Enums;
 using Cmf.CLI.Core.Objects;
 using Cmf.CLI.Utilities;
+using System.Collections.Generic;
+using System.CommandLine;
+using System.CommandLine.NamingConventionBinder;
+using System.IO.Abstractions;
+using System.Linq;
+using System;
 
 namespace Cmf.CLI.Commands.New
 {
@@ -51,7 +51,10 @@ namespace Cmf.CLI.Commands.New
         /// <param name="version">the package version</param>
         public void Execute(string version)
         {
-            var packageName = "Cmf.Custom.Tests";
+            var (organization, product) = GetOrganizationAndProductFromProjectConfig();
+
+            var packageName = $"{organization}.{product}.Tests";
+
             var projectRoot = FileSystemUtilities.GetProjectRoot(this.fileSystem);
             var repoType = ExecutionContext.Instance.ProjectConfig.RepositoryType ?? CliConstants.DefaultRepositoryType;
             var projectConfig = ExecutionContext.Instance.ProjectConfig;
@@ -75,6 +78,7 @@ namespace Cmf.CLI.Commands.New
             var vmHostname = projectConfig.vmHostname;
             var testScenariosNugetVersion = projectConfig.TestScenariosNugetVersion;
             var isSslEnabled = projectConfig.IsSslEnabled;
+
             args.AddRange(new[]
             {
                 "--projectName", projectName,
@@ -83,7 +87,9 @@ namespace Cmf.CLI.Commands.New
                 "--RESTPort", restPort.ToString(),
                 "--testScenariosNugetVersion", testScenariosNugetVersion.ToString(),
                 "--HTMLPort", htmlPort.ToString(),
-                "--MESVersion", mesVersion.ToString()
+                "--MESVersion", mesVersion.ToString(),
+                "--Product", product,
+                "--Organization", organization
             });
 
             if (isSslEnabled)
@@ -93,6 +99,11 @@ namespace Cmf.CLI.Commands.New
             
             #region version-specific bits
             args.AddRange(new []{ "--targetFramework", mesVersion.Major > 8 ? mesVersion.Major >= 11 ? "net8.0" : "net6.0" : "netcoreapp3.1" });
+
+            if (mesVersion >= new Version(11, 2, 3))
+            {
+                args.Add("--hostPerformanceTests");
+            }
             #endregion
 
             this.executedArgs = args.ToArray();