+++ cmf-cli/resources/template_feed/business/from900/Business.Package/%Org%.%Product%.Common/%Tenant%Utilities.cs
@@ -1,0 +1,6 @@
+﻿namespace <%= $CLI_PARAM_Organization %>.<%= $CLI_PARAM_Product %>.<%= $CLI_PARAM_idSegment %>.Common
+{
+    public class <%= $CLI_PARAM_Tenant %>Utilities
+    {
+    }
+}