+++ cmf-cli/resources/template_feed/business/from900/Business.Package/%Org%.%Product%.Orchestration/%Tenant%Orchestration.cs
@@ -1,0 +1,8 @@
+﻿
+namespace <%= $CLI_PARAM_Organization %>.<%= $CLI_PARAM_Product %>.<%= $CLI_PARAM_idSegment %>.Orchestration
+{
+    public class <%= $CLI_PARAM_Tenant %>Orchestration : I<%= $CLI_PARAM_Tenant %>Orchestration
+    {
+        private const string OBJECT_TYPE_NAME = "<%= $CLI_PARAM_Organization %>.<%= $CLI_PARAM_Product %>.<%= $CLI_PARAM_idSegment %>.Orchestration.<%= $CLI_PARAM_Tenant %>ManagementOrchestration";
+    }
+}