+++ cmf-cli/resources/template_feed/business/upto8x/Business.Package/%Org%.%Product%.Orchestration/%Tenant%Orchestration.cs
@@ -1,0 +1,7 @@
+﻿namespace <%= $CLI_PARAM_Organization %>.<%= $CLI_PARAM_Product %>.<%= $CLI_PARAM_idSegment %>.Orchestration
+{
+    public static class <%= $CLI_PARAM_Tenant %>Orchestration
+    {
+        private const string OBJECT_TYPE_NAME = "<%= $CLI_PARAM_Organization %>.<%= $CLI_PARAM_Product %>.<%= $CLI_PARAM_idSegment %>.Orchestration.<%= $CLI_PARAM_Tenant %>ManagementOrchestration";
+    }
+}